Moved Password to foam.util and updated references
